+++ classes/internal/Scoring/ScoringMetrics/FunctionalDependencies/class.FunctionalDependencies.php
@@ -195,7 +195,7 @@ class FunctionalDependencies extends ScoringMetric
         }
 
         // Compute 1 - Jaccard distance
-        return (1 - ((sizeof($union) - sizeof($intersect))/sizeof($union))) * $solution_metric->getPoints();
+        return (int)(1 - ((sizeof($union) - sizeof($intersect))/sizeof($union))) * $solution_metric->getPoints();
     }
 
     /**
